Deduplicate VT types, drop HistoryTrack re-export, and map agent surfaces.
Keep one react-view-transition ambient file in packages/ui, import HistoryTrack from services, group dev-console status UI, and document agent discovery plus Flags vs secrets.
