--- apps/web/src/app/dev-console/StatusChip.tsx
@@ -1,13 +1,0 @@
-import { Chip } from '@mui/material';
-
-/**
- * Status chip showing connected/not connected state.
- */
-export function StatusChip({ isConnected }: { isConnected: boolean }) {
-  return (
-    <Chip
-      color={isConnected ? 'success' : 'default'}
-      label={isConnected ? 'Connected' : 'Not connected'}
-    />
-  );
-}
