+++ apps/web/src/app/dev-console/oauth/OauthCard.tsx
@@ -7,8 +7,7 @@ import { Skeleton, Stack, Typography } from '@mui/material';
 import { connection } from 'next/server';
 import { Suspense } from 'react';
 import { DevConsoleCardShell } from '../DevConsoleCardShell';
-import { ErrorMessage } from '../ErrorMessage';
-import { StatusChip } from '../StatusChip';
+import { ErrorMessage, StatusChip } from '../StatusIndicators';
 import { ConnectButton } from './ConnectButton';
 import { ForceRefreshButton } from './ForceRefreshButton';
 import { TokenExpiry } from './TokenExpiry';
