+++ apps/web/src/app/dev-console/webhooks/WebhookCard.tsx
@@ -6,8 +6,7 @@ import { Skeleton, Stack, Typography } from '@mui/material';
 import { connection } from 'next/server';
 import { Suspense } from 'react';
 import { DevConsoleCardShell } from '../DevConsoleCardShell';
-import { ErrorMessage } from '../ErrorMessage';
-import { StatusChip } from '../StatusChip';
+import { ErrorMessage, StatusChip } from '../StatusIndicators';
 import { CreateWebhookButton } from './CreateWebhookButton';
 import { DeleteWebhookButton } from './DeleteWebhookButton';
 import { SubscriptionDetails } from './SubscriptionDetails';
