--- apps/web/src/types/react-view-transition.d.ts
@@ -1,22 +1,0 @@
-import type { ExoticComponent, ReactNode } from 'react';
-
-type ViewTransitionClass =
-  | 'none'
-  | 'auto'
-  | (string & {})
-  | Record<string, 'none' | 'auto' | string>;
-
-declare module 'react' {
-  interface ViewTransitionProps {
-    children?: ReactNode;
-    default?: ViewTransitionClass;
-    enter?: ViewTransitionClass;
-    exit?: ViewTransitionClass;
-    name?: string;
-    share?: ViewTransitionClass;
-    update?: ViewTransitionClass;
-  }
-
-  const ViewTransition: ExoticComponent<ViewTransitionProps>;
-  function addTransitionType(type: string): void;
-}
